% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_otu.R
\name{compare_observed_otus_multiple}
\alias{compare_observed_otus_multiple}
\title{Compare Observed OTUs Between Multiple Groups}
\usage{
compare_observed_otus_multiple(abundance_matrix)
}
\arguments{
\item{abundance_matrix}{Matrix where rows are species and columns are samples}
}
\value{
Matrix of p-values for pairwise comparisons
}
\description{
Compare Observed OTUs Between Multiple Groups
}
\examples{
communities <- matrix(c(10,20,0,40, 40,30,20,0, 25,0,25,25), 
                     nrow=4, ncol=3)
compare_observed_otus_multiple(communities)
}
\author{
Akinsuyi Oluwamayowa Samuel
}
