% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_otu.R
\name{compare_observed_otus_two}
\alias{compare_observed_otus_two}
\title{Compare Observed OTUs Between Two Groups}
\usage{
compare_observed_otus_two(x, y)
}
\arguments{
\item{x}{Numeric vector of abundances for first community}

\item{y}{Numeric vector of abundances for second community}
}
\value{
List containing test statistics and p-value
}
\description{
Compare Observed OTUs Between Two Groups
}
\examples{
sample1 <- c(10, 20, 30, 0, 40)
sample2 <- c(40, 0, 30, 20, 10)
compare_observed_otus_two(sample1, sample2)
}
\author{
Akinsuyi Oluwamayowa Samuel
}
