% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpson_diversity.R
\name{compare_simpson_multiple}
\alias{compare_simpson_multiple}
\title{Compare Simpson Diversity Between Multiple Groups}
\usage{
compare_simpson_multiple(abundance_matrix)
}
\arguments{
\item{abundance_matrix}{Matrix where rows are species and columns are samples}
}
\value{
Matrix of p-values for pairwise comparisons
}
\description{
Compare Simpson Diversity Between Multiple Groups
}
\examples{
communities <- matrix(c(10,20,30,40, 40,30,20,10, 25,25,25,25), 
                     nrow=4, ncol=3)
compare_simpson_multiple(communities)
}
\author{
Akinsuyi Oluwamayowa Samuel
}
