% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpson_diversity.R
\name{compare_simpson_two}
\alias{compare_simpson_two}
\title{Compare Simpson Diversity Between Two Groups}
\usage{
compare_simpson_two(x, y)
}
\arguments{
\item{x}{Numeric vector of abundances for first community}

\item{y}{Numeric vector of abundances for second community}
}
\value{
List containing test statistics and p-value
}
\description{
Compare Simpson Diversity Between Two Groups
}
\examples{
sample1 <- c(10, 20, 30, 40)
sample2 <- c(40, 30, 20, 10)
compare_simpson_two(sample1, sample2)
}
\author{
Akinsuyi Oluwamayowa Samuel
}
