% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microbialdiversity-package.R
\name{Microbialdiversity}
\alias{Microbialdiversity}
\title{Microbialdiversity: A package for calculating microbial diversity indices}
\description{
Microbialdiversity: A package for calculating microbial diversity indices
}
